CI: Get repo back to green (#189)
#### Problem

The token-wrap repo is currently failing CI because of some changes that
went in without passing all the checks.

This was my fault, because I didn't add the new checks to the required
list in order to merge.

#### Summary of changes

All of the steps are now required on the GitHub repo, but there's still
some fixes to get to green.

* Downgrade to kit v2.2.1. This is needed because codama is generating
  code with IInstruction, which has been deprecated
* Build programs individually in the `programs:build` step. Currently
  this is also trying to build the CLI, which fails.
* I re-ran `pnpm generate` to get things up to speed.